requirements.txt
pandas==1.4.4
requests==2.28.1
plotly==5.10.0
streamlit==1.13.0
xlrd >= 1.0.0
lxml==4.9.1
openpyxl==3.0.10
python-dotenv==0.19.2